+++ lib/WorkflowMain.groovy
@@ -46,6 +46,7 @@ class WorkflowMain {
     // Validate parameters and print summary to screen
     //
     public static void initialise(workflow, params, log) {
+
         // Print help to screen if required
         if (params.help) {
             log.info help(workflow, params, log)
@@ -60,6 +61,9 @@ class WorkflowMain {
         // Print parameter summary log to screen
         log.info paramsSummaryLog(workflow, params, log)
 
+        // Check that a -profile or Nextflow config has been provided to run the pipeline
+        NfcoreTemplate.checkConfigProvided(workflow, log)
+
         // Check that conda channels are set-up correctly
         if (params.enable_conda) {
             Utils.checkCondaChannels(log)
@@ -68,13 +72,67 @@ class WorkflowMain {
         // Check AWS batch settings
         NfcoreTemplate.awsBatch(workflow, params)
 
-        // Check the hostnames against configured profiles
-        NfcoreTemplate.hostName(workflow, params, log)
-
         // Check input has been provided
         if (!params.input) {
-            log.error "Please provide an input samplesheet to the pipeline e.g. '--input ids.txt'"
+            log.error "Please provide an input file containing ids to the pipeline - one per line e.g. '--input ids.txt'"
             System.exit(1)
         }
+
+        // Check valid input_type has been provided
+        def input_types = ['sra', 'synapse']
+        if (!input_types.contains(params.input_type)) {
+            log.error "Invalid option: '${params.input_type}'. Valid options for '--input_type': ${input_types.join(', ')}."
+            System.exit(1)
+        }
+    }
+
+    // Check if input ids are from the SRA
+    public static Boolean isSraId(input, log) {
+        def is_sra = false
+        def total_ids = 0
+        def no_match_ids = []
+        def pattern = /^(((SR|ER|DR)[APRSX])|(SAM(N|EA|D))|(PRJ(NA|EB|DB))|(GS[EM]))(\d+)$/
+        input.eachLine { line ->
+            total_ids += 1
+            if (!(line =~ pattern)) {
+                no_match_ids << line
+            }
+        }
+
+        def num_match = total_ids - no_match_ids.size()
+        if (num_match > 0) {
+            if (num_match == total_ids) {
+                is_sra = true
+            } else {
+                log.error "Mixture of ids provided via --input: ${no_match_ids.join(', ')}\nPlease provide either SRA / ENA / DDBJ / GEO or Synapse ids!"
+                System.exit(1)
+            }
+        }
+        return is_sra
+    }
+
+    // Check if input ids are from the Synapse platform
+    public static Boolean isSynapseId(input, log) {
+        def is_synapse = false
+        def total_ids = 0
+        def no_match_ids = []
+        def pattern = /^syn\d{8}$/
+        input.eachLine { line ->
+            total_ids += 1
+            if (!(line =~ pattern)) {
+                no_match_ids << line
+            }
+        }
+
+        def num_match = total_ids - no_match_ids.size()
+        if (num_match > 0) {
+            if (num_match == total_ids) {
+                is_synapse = true
+            } else {
+                log.error "Mixture of ids provided via --input: ${no_match_ids.join(', ')}\nPlease provide either SRA / ENA / DDBJ / GEO or Synapse ids!"
+                System.exit(1)
+            }
+        }
+        return is_synapse
     }
 }