+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [[1.4](https://github.com/nf-core/fetchngs/releases/tag/1.4)] - 2021-11-09
+
+### Enhancements & fixes
+
+* Convert pipeline to updated Nextflow DSL2 syntax for future adoption across nf-core
+* Added a workflow to download FastQ files and to create samplesheets for ids from the [Synapse platform](https://www.synapse.org/) hosted by [Sage Bionetworks](https://sagebionetworks.org/).
+* SRA identifiers not available for direct download via the ENA FTP will now be downloaded via [`sra-tools`](https://github.com/ncbi/sra-tools).
+* Added `--force_sratools_download` parameter to preferentially download all FastQ files via `sra-tools` instead of ENA FTP.
+* Correctly handle errors from SRA identifiers that do **not** return metadata, for example, due to being private.
+* Retry an error in prefetch via bash script in order to allow it to resume interrupted downloads.
+* Name output FastQ files by `{EXP_ACC}_{RUN_ACC}*fastq.gz` instead of `{EXP_ACC}_{T*}*fastq.gz` for run id provenance
+* [[#46](https://github.com/nf-core/fetchngs/issues/46)] - Bug in sra_ids_to_runinfo.py
+* Added support for [DDBJ ids](https://www.ddbj.nig.ac.jp/index-e.html). See examples below:
+
+| `DDBJ`        |
+|---------------|
+| PRJDB4176     |
+| SAMD00114846  |
+| DRA008156     |
+| DRP004793     |
+| DRR171822     |
+| DRS090921     |
+| DRX162434     |
+
 ## [[1.3](https://github.com/nf-core/fetchngs/releases/tag/1.3)] - 2021-09-15
 
 ### Enhancements & fixes